# docs/swagger/swagger.yaml
definitions:
...
  http.standardJsonResponse:
    properties:
      data:
        type: object
      error:
        type: string
      meta:
        type: object
      success:
        type: boolean
    type: object
...
  http.subscriptionAssociateRequestBody:
    properties:
      organization_id:
        type: string
      team_id:
        type: string
    type: object
  http.subscriptionRequestBody:
    properties:
      active:
        type: boolean
      customer_id:
        type: string
      expires:
        type: string
      organization_id:
        type: string
      subscription_id:
        type: string
      team_id:
        type: string
      type:
        enum:
        - user
        - ' team'
        - ' organization'
        type: string
      user_id:
        type: string
    required:
    - type
    type: object
...
  thunderdome.Subscription:
    properties:
      active:
        type: boolean
      created_date:
        type: string
      customer_id:
        type: string
      expires:
        type: string
      id:
        type: string
      organization_id:
        type: string
      subscription_id:
        type: string
      team_id:
        type: string
      type:
        type: string
      updated_date:
        type: string
      user:
        $ref: '#/definitions/thunderdome.User'
      user_id:
        type: string
    type: object
...
  thunderdome.User:
    properties:
      avatar:
        type: string
      company:
        type: string
      country:
        type: string
      createdDate:
        type: string
      disabled:
        type: boolean
      email:
        type: string
      gravatarHash:
        type: string
      id:
        type: string
      jobTitle:
        type: string
      lastActive:
        type: string
      locale:
        type: string
      name:
        type: string
      notificationsEnabled:
        type: boolean
      picture:
        type: string
      rank:
        type: string
      theme:
        type: string
      updatedDate:
        type: string
      verified:
        type: boolean
    type: object
...
info:
  contact:
    name: Steven Weathers
    url: https://github.com/StevenWeathers/thunderdome-planning-poker
  description: |-
    Thunderdome Planning Poker API for both Internal and External use.
    WARNING: Currently not considered stable and is subject to change until 1.0 is released.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Thunderdome API
  version: BETA
paths:
...
  /subscriptions:
    get:
      description: get list of subscriptions
      parameters:
      - description: Max number of results to return
        in: query
        name: limit
        type: integer
      - description: Starting point to return rows from, should be multiplied by limit
          or 0
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/thunderdome.Subscription'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Subscriptions
      tags:
      - subscription
    post:
      description: Creates a subscription
      parameters:
      - description: new subscription object
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/http.subscriptionRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/thunderdome.Subscription'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Subscription
      tags:
      - subscription
  /subscriptions/{subscriptionId}:
    delete:
      description: Deletes a Subscription
      parameters:
      - description: the subscription ID to delete
        in: path
        name: subscriptionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Subscription
      tags:
      - subscription
    get:
      description: Get a subscription
      parameters:
      - description: the subscription ID
        in: path
        name: subscriptionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/thunderdome.Subscription'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Subscription
      tags:
      - subscription
    put:
      description: Updates a Subscription
      parameters:
      - description: the subscription ID to update
        in: path
        name: subscriptionId
        required: true
        type: string
      - description: subscription object to update
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/http.subscriptionRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/thunderdome.Subscription'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Subscription
      tags:
      - subscription
...
  /users/{userId}/subscriptions:
    get:
      description: get list of active entity user subscriptions
      parameters:
      - description: the entity user ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/thunderdome.Subscription'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Entity User Active Subscriptions
      tags:
      - subscription
  /users/{userId}/subscriptions/{subscriptionId}:
    patch:
      description: get list of active entity user subscriptions
      parameters:
      - description: the entity user ID
        in: path
        name: userId
        required: true
        type: string
      - description: the subscription ID to update
        in: path
        name: subscriptionId
        required: true
        type: string
      - description: update subscription association object
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/http.subscriptionAssociateRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/thunderdome.Subscription'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Entity User Subscriptions
      tags:
      - subscription
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"